+++ src/Oci8/Oci8Connection.php
@@ -132,6 +132,34 @@ class Oci8Connection extends Connection
     }
 
     /**
+     * Get configured Oracle server version.
+     */
+    public function serverVersion(): string
+    {
+        return $this->getConfig('server_version') ?? '11g';
+    }
+
+    public function isVersionAbove(string $version): bool
+    {
+        return version_compare($version, $this->serverVersion(), '<');
+    }
+
+    public function isVersionAboveOrEqual(string $version): bool
+    {
+        return version_compare($version, $this->serverVersion(), '<=');
+    }
+
+    public function isVersionBelow(string $version): bool
+    {
+        return version_compare($version, $this->serverVersion(), '>');
+    }
+
+    public function isVersionBelowOrEqual(string $version): bool
+    {
+        return version_compare($version, $this->serverVersion(), '>=');
+    }
+
+    /**
      * Get a new query builder instance.
      */
     public function query(): QueryBuilder
@@ -152,6 +180,16 @@ class Oci8Connection extends Connection
         ];
 
         return $this->setSessionVars($sessionVars);
+    }
+
+    /**
+     * Get the configured NLS_DATE_FORMAT from session variables.
+     */
+    public function getDateFormat(): string
+    {
+        $sessionVars = $this->getConfig('sessionVars') ?? [];
+
+        return $sessionVars['NLS_DATE_FORMAT'] ?? 'YYYY-MM-DD HH24:MI:SS';
     }
 
     /**