+++ tests/Functional/QueryBuilderTest.php
@@ -3,7 +3,7 @@
 namespace Yajra\Oci8\Tests\Functional;
 
 use Illuminate\Database\Query\Expression as Raw;
-use Illuminate\Foundation\Testing\DatabaseTransactions;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use PHPUnit\Framework\Attributes\Test;
 use Yajra\Oci8\Query\Grammars\OracleGrammar;
@@ -11,10 +11,61 @@ use Yajra\Oci8\Query\OracleBuilder as Builder;
 use Yajra\Oci8\Query\Processors\OracleProcessor;
 use Yajra\Oci8\Schema\OracleBlueprint as Blueprint;
 use Yajra\Oci8\Tests\TestCase;
+use Yajra\Oci8\Tests\User;
 
 class QueryBuilderTest extends TestCase
 {
-    use DatabaseTransactions;
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        Schema::create('users', function (\Illuminate\Database\Schema\Blueprint $table) {
+            $table->increments('id');
+            $table->string('name');
+            $table->string('email');
+            $table->timestamps();
+        });
+
+        Schema::create('jobs', function (\Illuminate\Database\Schema\Blueprint $table) {
+            $table->increments('id');
+            $table->string('name')->nullable();
+            $table->integer('job_id')->nullable();
+        });
+
+        Schema::create('empty_defaults_table', function (Blueprint $table) {
+            $table->id('id');
+            $table->string('name')->default('default');
+        });
+
+        Schema::create('multiple_raw_insert_table', function (Blueprint $table) {
+            $table->id('id');
+            $table->string('name');
+        });
+
+        Schema::create('single_raw_insert_table', function (Blueprint $table) {
+            $table->id('id');
+            $table->string('name');
+        });
+
+        collect(range(1, 20))->each(function ($i) {
+            /** @var User $user */
+            User::query()->create([
+                'name' => 'Record-'.$i,
+                'email' => 'Email-'.$i.'@example.com',
+            ]);
+        });
+    }
+
+    protected function tearDown(): void
+    {
+        Schema::drop('users');
+        Schema::drop('jobs');
+        Schema::drop('empty_defaults_table');
+        Schema::drop('multiple_raw_insert_table');
+        Schema::drop('single_raw_insert_table');
+
+        parent::tearDown();
+    }
 
     #[Test]
     public function it_can_perform_insert()
@@ -70,15 +121,6 @@ class QueryBuilderTest extends TestCase
     #[Test]
     public function it_can_insert_empty_and_get_id()
     {
-        if (Schema::hasTable('empty_defaults_table')) {
-            Schema::drop('empty_defaults_table');
-        }
-
-        Schema::create('empty_defaults_table', function (Blueprint $table) {
-            $table->id('id');
-            $table->string('name')->default('default');
-        });
-
         $id = $this->getBuilder()->from('empty_defaults_table')->insertGetId([]);
 
         $this->assertSame(1, $id);
@@ -91,15 +133,6 @@ class QueryBuilderTest extends TestCase
     #[Test]
     public function it_can_insert_single_using_raw_query()
     {
-        if (Schema::hasTable('single_raw_insert_table')) {
-            Schema::drop('single_raw_insert_table');
-        }
-
-        Schema::create('single_raw_insert_table', function (Blueprint $table) {
-            $table->id('id');
-            $table->string('name');
-        });
-
         $this->getBuilder()->from('single_raw_insert_table')->insert([
             ['name' => new Raw("UPPER('Foo')")],
         ]);
@@ -113,15 +146,6 @@ class QueryBuilderTest extends TestCase
     #[Test]
     public function it_can_insert_multiple_using_raw_query()
     {
-        if (Schema::hasTable('multiple_raw_insert_table')) {
-            Schema::drop('multiple_raw_insert_table');
-        }
-
-        Schema::create('multiple_raw_insert_table', function (Blueprint $table) {
-            $table->id('id');
-            $table->string('name');
-        });
-
         $this->getBuilder()->from('multiple_raw_insert_table')->insert([
             ['name' => new Raw("UPPER('Foo')")],
             ['name' => new Raw("LOWER('Foo')")],
@@ -135,6 +159,23 @@ class QueryBuilderTest extends TestCase
             'id' => 2,
             'name' => 'foo',
         ]);
+    }
+
+    #[Test]
+    public function it_keeps_rn_in_pagination_when_selected()
+    {
+        $expected = User::select(['name', DB::raw('ROWNUM rn')])->limit(2)->orderBy('id')->get()->toArray();
+        $notexpected = User::select(['name'])->limit(2)->orderBy('id')->get()->toArray();
+        $notexpected2 = User::select(['name'])->limit(2)->orderBy('id')->get()->toArray();
+
+        $this->assertArrayHasKey('rn', $expected[0]);
+        $this->assertArrayHasKey('rn', $expected[1]);
+
+        $this->assertArrayNotHasKey('rn', $notexpected[0]);
+        $this->assertArrayNotHasKey('rn', $notexpected[1]);
+
+        $this->assertArrayNotHasKey('rn', $notexpected2[0]);
+        $this->assertArrayNotHasKey('rn', $notexpected2[1]);
     }
 
     protected function getBuilder(): Builder